refactor: 重新封装邀请成员按钮
