+++ rollingdesign/src/views/ProjectView.vue
@@ -28,29 +28,7 @@
                                 <el-button type="primary" @click="designDialogVisible = true">新建原型</el-button>
                                 <el-button type="primary" @click="newDoc()">新建文档</el-button>
                                 <!--如果是管理员有“邀请”这一项，判断登陆者在该团队中的身份-->
-                                <!--
-                                <el-button type="primary" @click="centerDialogVisible = true">邀请成员</el-button>
-                                -->
-                                <el-popover :width="300" trigger="click" ref='popper'
-                                    popper-style="box-shadow: rgb(14 18 22 / 35%) 0px 10px 38px -10px, rgb(14 18 22 / 20%) 0px 10px 20px -15px; padding: 20px;">
-                                    <template #reference>
-                                        <el-button type="primary" @click="generateLink()">邀请成员</el-button>
-                                        <!-- <el-avatar src="https://avatars.githubusercontent.com/u/72015883?v=4" /> -->
-                                    </template>
-                                    <template #default>
-                                        <div disabled style="text-align: center;">
-                                            <el-input v-model="link" disabled></el-input>
-                                            <div style="color: #d2d3d7;text-align: left;
-                            margin-top: 20px;">该链接将在<span style="font-weight: 700;">24小时</span>内过期
-                                            </div>
-                                        </div>
-                                        <button @click="copyLink" class="copyLink"
-                                            style="text-align: center; 
-                            background-color: #3671ff;
-                            outline: none;
-                        margin-left:auto; margin-right: auto; margin-top: 15px; padding: 10px; box-sizing: content-box;">复制链接</button>
-                                    </template>
-                                </el-popover>
+                                <InviteMemberButton :now-team="nowTeam" />
                             </div>
                         </el-row>
 
@@ -196,7 +174,7 @@
         </el-container>
     </div>
 </template>
-  
+
 <script setup>
 import qs from 'qs'
 import axios from 'axios'
@@ -230,6 +208,7 @@ const designNameInput = ref('')
 /*头像*/
 import { reactive, toRefs } from 'vue'
 import router from '@/router';
+import InviteMemberButton from "@/components/InviteMemberButton.vue";
 
 const jumpToDoc = (id) => {
     //this.$router.push('/video/'+video_id);
@@ -257,21 +236,6 @@ const jumpToDesign = (id) => {
 // import { ref } from 'vue'
 const centerDialogVisible = ref(false) /*邀请对话框*/
 const input = ref('')/*邀请成员时，输入框*/
-
-const link = ref('')
-const generateLink = async () => {
-
-    let Headers = { 'Authorization': authStore().token }
-
-    let res = await axios.get('http://www.aamofe.top/api/team/get_invitation/', {
-        headers: Headers,
-        params: {
-            team_id: nowTeam.teamId,
-        }
-    })
-    console.log(res.data)
-    link.value = res.data.invatation
-}
 
 /*main*/
 const designShow = ref(true)
@@ -438,7 +402,7 @@ const newDesign = () => {
     let formData = new FormData();
     formData.append("team_name", addTeamIntroductionInput.value);
     formData.append("Authorization", authStore().token);
-    // nowProject.projectId 
+    // nowProject.projectId
 
     axios.post('http://www.aamofe.top/api/team/create_project/' + nowProject.projectId + '/', qs.stringify({
         project_name: designNameInput.value
@@ -512,7 +476,7 @@ const newDoc = () => {
         });
 }
 </script>
-  
+
 <style scoped>
 .hintText {
     color: gray;
